pyproject.toml
[tool.poetry]
name = "streamlit-agent"
version = "0.1.0"
description = ""
authors = ["Tim Conkling <tconkling@gmail.com>", "Joshua Carroll <joshua.carroll@snowflake.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "streamlit_agent"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
langchain = {version = ">=0.0.252"}
openai = "^0.27.8"
duckduckgo-search = "^3.8.3"
pypdf = "^3.12.2"
sentence-transformers = "^2.2.2"
torch = ">=2.0.0, !=2.0.1"
tabulate = "^0.9.0"
streamlit-feedback = "^0.0.9"
langchain-experimental = "^0.0.10"
streamlit = ">=1.26"
docarray = "^0.38.0"
hnswlib = "^0.7.0"
python-dotenv = "^1.0.0"
pypdf2 = "^3.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.4.1"
pre-commit = "^3.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
